// src/app/models/empleado.interface.ts
export interface empleado {
    id_empleado?: number;
    id_direccion?: number;
    id_genero?: number;
    id_estado_civil?: number;
    primer_nombre?: string;
    segundo_nombre?: string;
    primer_apellido?: string;
    segundo_apellido?: string;
    DPI?: string;
    fecha_nacimiento?: Date;
    telefono?: number;
    correo_electronico?: string;
    fecha_contratacion?: Date;
  }